pyproject.toml
[project]
name = "ai-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.0",
    "fastapi>=0.123.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "numpy>=2.3.5",
    "pdfminer-six>=20251107",
    "pinecone-client>=6.0.0",
    "pydantic>=2.12.5",
    "pypdf>=6.4.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "tavily-python>=0.7.13",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.38.0",
    "wikipedia>=1.4.0",
]
